+++ src/armour.hpp
@@ -3,6 +3,7 @@
 
 #include "item.hpp"
 
+/* template used to describe armour items */
 struct ArmourTemplate {
     uint type;
     uint limb;
@@ -11,6 +12,7 @@ struct ArmourTemplate {
     uint AC;
 };
 
+/* limbs armour can be for */
 enum ArmourLimbs {
     ARMOUR_HELMET,
     ARMOUR_TORSO,
@@ -21,16 +23,18 @@ enum ArmourLimbs {
 };
 
 class Armour : public Item {
-    uint _limb;
-    uint _AC;
+    uint _limb; // what limb this armour is worn on
+    uint _AC;   // armour class
 
 public:
+    /* constructor takes an ArmourTemplate */
     Armour(const ArmourTemplate& at) :
         Item(at.name, ITEM_ARMOUR, at.type, at.weight),
         _limb(at.limb),
         _AC(at.AC)
     {}
 
+    /* get methods for attributes */
     uint Limb() const {return _limb;}
     uint AC() const {return _AC;}
 };