+++ src/character.hpp
@@ -4,6 +4,7 @@
 #include "common.hpp"
 #include "creature.hpp"
 
+/* enumeration of all available races in the game */
 enum Races {
     RACE_HUMAN,
     RACE_ELF,
@@ -11,14 +12,9 @@ enum Races {
     RACE_HALFLING,
     RACE_GNOME,
     RACE_HALF_ORC,
-    RACE_DROMITE,
-    RACE_DUERGAR,
-    RACE_ELAN,
-    RACE_HALF_GIANT,
-    RACE_MAENAD,
-    RACE_XEPH
 };
 
+/* enumeration of all available classes in the game */
 enum Classes {
     CLASS_BARBARIAN,
     CLASS_BARD,
@@ -27,16 +23,13 @@ enum Classes {
     CLASS_FIGHTER,
     CLASS_MONK,
     CLASS_PALADIN,
-    CLASS_PSION,
-    CLASS_PSYCHIC_WARRIOR,
     CLASS_RANGER,
     CLASS_ROGUE,
     CLASS_SORCEROR,
-    CLASS_SOULKNIFE,
-    CLASS_WILDER,
     CLASS_WIZARD
 };
 
+/* enumeration of all available alignments in the game */
 enum Alignments {
     LAWFUL_GOOD,
     LAWFUL_NEUTRAL,
@@ -50,24 +43,29 @@ enum Alignments {
 };
 
 class Character : public Creature {
-    std::string _name;
-    uint _age;
-    bool _sex;
-    uchar _race;
-    uchar __class;
-    uchar _alignment;
+    std::string _name;  // name
+    uint _age;          // age
+    bool _sex;          // gender
+    uchar _race;        // race
+    uchar __class;      // class
+    uchar _alignment;   // alignment
 
 public:
+    /*
+     * Character constructor takes arguments to set the above attributes.
+     * Only name is required, default values are provided for the others
+     */
     Character(std::string name, uint age = 30, bool sex = true,
             uchar race = RACE_HUMAN, uchar _class = CLASS_FIGHTER,
             uchar alignment = TRUE_NEUTRAL);
 
-    std::string Name() const {return _name;};
-    uint Age() const {return _age;};
-    bool Sex() const {return _sex;};
-    uchar Race() const {return _race;};
-    uchar Class() const {return __class;};
-    uchar Alignment() const {return _alignment;};
+    /* get methods for attributes */
+    std::string Name() const {return _name;}
+    uint Age() const {return _age;}
+    bool Sex() const {return _sex;}
+    uchar Race() const {return _race;}
+    uchar Class() const {return __class;}
+    uchar Alignment() const {return _alignment;}
 };
 
 #endif