+++ src/common.hpp
@@ -1,6 +1,10 @@
 #ifndef _LoX_COMMON
 #define _LoX_COMMON
 
+/* version string */
+#define _LoX_VERSION 0.2.0.0
+
+/* global external headers */
 #include <cstdio>
 #include <cstdlib>
 #include <cstring>
@@ -11,15 +15,19 @@
 #include <vector>
 #include <list>
 
+/* global internal headers */
 #include "dieroll.hpp"
 
+/* easier declaration of unsigned integers */
 typedef unsigned char uchar;
 typedef unsigned int uint;
 
+/* this structure is used to describe 2D positions and movements in the game */
 struct Vector2D {
     uint x;
     uint y;
 
+    /* overload for '==' operator to work correctly */
     bool operator==(const Vector2D& vect) {
         if(x == vect.x && y == vect.y) return true;
         return false;