+++ src/weapon.hpp
@@ -3,6 +3,7 @@
 
 #include "item.hpp"
 
+/* template used to describe weapon items */
 struct WeaponTemplate {
     uint type;
     uint style;
@@ -12,6 +13,7 @@ struct WeaponTemplate {
     uint crit;
 };
 
+/* all available weapon styles */
 enum WeaponStyles {
     WEAPON_SLASHING,
     WEAPON_PIERCING,
@@ -19,11 +21,12 @@ enum WeaponStyles {
 };
 
 class Weapon : public Item {
-    uint _style;
-    Die _dmg;
-    uint _crit;
+    uint _style;    // weapon style
+    Die _dmg;       // hit die
+    uint _crit;     // critical multiplier
 
 public:
+    /* constructor takes a WeaponTemplate */
     Weapon(const WeaponTemplate& wt) :
         Item(wt.name, ITEM_WEAPON, wt.type, wt.weight),
         _style(wt.style),
@@ -31,6 +34,7 @@ public:
         _crit(wt.crit)
     {}
 
+    /* get methods for attributes */
     uint Style() const {return _style;}
     Die Dmg() const {return _dmg;}
     uint Crit() const {return _crit;}